ConsoleUi: added some methods to tihs class which replace the messages got printed inside the TodoService so SRP must be followed !

+++ TodoApp/TodoApp/service/TodoService.cs
@@ -46,13 +46,18 @@ namespace TodoApp.service
             {
                 repo.SeeAllToDos();
             }
-            else if (input.Equals("A")) {
+            else if (input.Equals("A")) 
+            {
  
                 repo.AddToDo();
             }
-            else // this part gets R
+            else if (input.Equals("R")) 
             {
                 repo.RemoveToDo();
+            }
+            else
+            {
+                return;
             }
